Add compilePackingList, compileSubmissionData to utilities
